Fix fractional scaling implementation
Previously the fractional scaling implementation would send size
requests based on the integer scale to all clients. This however is
incorrect.

Instead both fractional and traditional clients now receive the size
request at the fractional logical scale. Fractional clients use this to
render the buffer at the correct size immediately, while traditional
clients render it at a scale of 2 and then get downscaled by the
compositor.

+++ src/windows/layout.rs
@@ -398,7 +398,7 @@ impl Layouts {
     }
 
     /// Add all visible windows' textures to the supplied buffer.
-    pub fn textures(&self, textures: &mut Vec<CatacombElement>, scale: i32) {
+    pub fn textures(&self, textures: &mut Vec<CatacombElement>, scale: f64) {
         let layout = self.active();
 
         if let Some(secondary) = layout.secondary().map(|window| window.borrow()) {
